finished rework on existing network architecture and conitnuing updating, thinking about moving dynamicclienthandler to generel clienthandler, merged some net message files to make access easier, operator as it is working

+++ Operator/source/database/KeyTable.h
@@ -86,7 +86,7 @@ namespace Database
 					statement,
 					columnIndex,
 					content.source.c_str(),
-					content.source.size(),
+					(int) content.source.size(),
 					SQLITE_STATIC);
 
 			case Column::Type:
